add __init__.py and change the file to right place
